+++ library/src/android/support/v4/view/accessibility/AccessibilityNodeInfoCompat.java
@@ -25,7 +25,8 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * Helper for accessing AccessibilityNodeInfo from newer platform versions.
+ * Helper for accessing {@link android.view.accessibility.AccessibilityNodeInfo}
+ * introduced after API level 4 in a backwards compatible fashion.
  */
 public class AccessibilityNodeInfoCompat {