+++ library/src/android/support/v4/view/accessibility/AccessibilityRecordCompat.java
@@ -24,7 +24,8 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * Helper for accessing AccessibilityRecord from newer platform versions.
+ * Helper for accessing {@link android.view.accessibility.AccessibilityRecord}
+ * introduced after API level 4 in a backwards compatible fashion.
  */
 public class AccessibilityRecordCompat {